Address review feedback on PR 98
Move unwrapCommandResult into services/command-result.ts so non-Daylite
services no longer import Daylite helpers; daylite-service-helpers.ts
keeps only normalizeOptionalString and the specs are split accordingly.
Guard useZepCalendars.ensureLoaded with an in-flight ref so two calls in
the same render cannot start two discoveries.
