+++ src/app/hooks/use-zep-calendars.ts
@@ -1,4 +1,4 @@
-import { useCallback, useState } from "react";
+import { useCallback, useRef, useState } from "react";
 import type { ZepCalendar } from "../../generated/tauri";
 import { discoverZepCalendars } from "../../services/zep";
 
@@ -16,9 +16,13 @@ export function useZepCalendars(): ZepCalendarsState {
   const [calendars, setCalendars] = useState<ZepCalendar[] | null>(null);
   const [isLoading, setIsLoading] = useState(false);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
+  // Guards ensureLoaded against double-fetching: the isLoading state only
+  // updates on the next render, the ref flips immediately.
+  const isFetching = useRef(false);
 
   const reload = useCallback(() => {
     void (async () => {
+      isFetching.current = true;
       setIsLoading(true);
       setErrorMessage(null);
       try {
@@ -30,16 +34,17 @@ export function useZepCalendars(): ZepCalendarsState {
             : "Die ZEP-Kalender konnten nicht geladen werden.",
         );
       } finally {
+        isFetching.current = false;
         setIsLoading(false);
       }
     })();
   }, []);
 
   const ensureLoaded = useCallback(() => {
-    if (calendars === null && !isLoading) {
+    if (calendars === null && !isFetching.current) {
       reload();
     }
-  }, [calendars, isLoading, reload]);
+  }, [calendars, reload]);
 
   return { calendars, isLoading, errorMessage, reload, ensureLoaded };
 }
