+++ src/services/display-settings.ts
@@ -1,5 +1,5 @@
 import { commands, type DisplaySettings } from "../generated/tauri";
-import { unwrapCommandResult } from "./daylite-service-helpers";
+import { unwrapCommandResult } from "./command-result";
 
 // Mirrors DisplaySettings::default() in the Rust backend.
 const defaultDisplaySettings: Required<DisplaySettings> = {
